+++ particle_simulations/configparser.h
@@ -23,7 +23,7 @@ struct Particle {
 
 struct Config {
     int num_time_step = 0;
-    int time_step_length = 0;
+    double time_step_length = 0;
     double sigma = 0.0f;
     double epsilon = 0.0f;
     int particle_num = 0;
